refactor/ configs method and moving Date() to LocalDateTime()
