+++ src/main/java/model/Car.java
@@ -1,6 +1,7 @@
+package model;
 
 public class Car implements Vehicle {
   public String getType() {
-    return "Car";
+    return "model.Car";
   }
 }
