+++ src/main/java/model/Vehicle.java
@@ -1,3 +1,4 @@
+package model;
 
 public interface Vehicle {
