+++ src/test/java/RushHoursHubTest.java
@@ -1,0 +1,2 @@
+public class RushHoursHubTest {
+}
